Updated logic
