+++ src/components/pages/Categories.jsx
