+++ src/index.js
@@ -5,7 +5,7 @@ import { Provider } from 'react-redux';
 import { BrowserRouter as Router } from 'react-router-dom';
 import './index.css';
 import App from './App';
-import store from './redux/configureStore';
+import store from './redux/configureStore/configureStore';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
