+++ src/redux/configureStore/configureStore.jsx
@@ -1,6 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit';
-import booksReducer from './books/books';
-import categoriesReducer from './categories/categories';
+import booksReducer from '../books/books';
+import categoriesReducer from '../categories/categories';
 
 const store = configureStore({
   reducer: {
